raoufcherfa/backend/requirements.txt
flask
flask-cors
cors
pytest
mysql-connector-python
mysql.connector
